[HOTSPOT-188] refactor: 구조 수정
