--- apps/web/src/entities/policy-add/index.ts
@@ -1,2 +1,0 @@
-export { BlockAddItem } from './ui/BlockAddItem';
-export { PolicyAddItem } from './ui/PolicyAddItem';
