--- apps/web/src/entities/policy-order/index.ts
@@ -1,2 +1,0 @@
-export type { FamilyPriority, PolicyOrderType } from './model/type';
-export { POLICY_ORDER_TYPE } from './model/type';
