+++ apps/web/src/entities/policy/index.ts
@@ -1,5 +1,14 @@
 export { policyDescriptionFormatter } from './lib/policyDescriptionFormatter';
-export type { BlockPolicy, Policy, PolicyPerFamily, PolicyPerUser } from './model/types';
+export type {
+  BlockPolicy,
+  FamilyPriority,
+  MemberPriority,
+  Policy,
+  PolicyOrderType,
+  PolicyPerFamily,
+  PolicyPerUser,
+} from './model/types';
+export { POLICY_ORDER_TYPE } from './model/types';
 export { useBlock } from './model/useBlock';
 export { useFamilyPolicy } from './model/useFamilyPolicy';
 export { usePolicy } from './model/usePolicy';
