+++ apps/web/src/features/policy-apply/ui/PolicyAddList.tsx
@@ -1,7 +1,7 @@
 import type { Policy } from '@entities/policy';
 import { policyDescriptionFormatter, usePolicy } from '@entities/policy';
-import { PolicyAddItem } from '@entities/policy-add';
 import type { PolicyApply } from '../model/types';
+import { PolicyAddItem } from './PolicyAddItem';
 
 interface PolicyAddListProps {
   data: Policy[];
